qlog: change event time field from f32 to f64
This change is a stepping stop to supporting
https://datatracker.ietf.org/doc/html/draft-ietf-quic-qlog-main-schema-13.
It increases the precision of the qlog event time field from
32-bit to 64-bit floating point.

While it would have been nice to do this as one atomic change,
it ends up touching a lot of dependent crates for such a small
technical change. Hence, spliting it out to make other qlog
update PRs easier to review.

Changes:
- Updated qlog Event and JsonEvent time fields from f32 to f64
- Updated Eventable trait set_time() method signature
- Updated Event::with_time() and with_time_ex() constructors
- Updated all time conversions using as_secs_f32() to as_secs_f64()
- Updated qlog-dancer types to use f64 for time coordinates
- RTT values remain as f32 since we aren't travelling to outerspace yet.

+++ h3i/src/recordreplay/qlog.rs
@@ -627,7 +627,7 @@ mod tests {
     use quiche::h3::Header;
     use serde_json;
 
-    const NOW: f32 = 123.0;
+    const NOW: f64 = 123.0;
     const H3I_WAIT: &str = "h3i:wait";
 
     #[test]
